requirements.txt
qdrant-client==1.9.1
pillow==10.3.0
numpy==1.26.4
sentence-transformers==2.7.0
torch>=2.0.0
torchvision>=0.15.0
tqdm==4.66.4
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
yt-dlp==2024.12.13
ffmpeg-python==0.2.0